차트 숫자 오버플로우 해결, 디테일페이지 매니저 아니면 쫒아내기

+++ src/main/java/kr/or/eum/manager/model/vo/Chart.java
@@ -18,8 +18,8 @@ public class Chart {
 	private int blackCount;		//블랙리스트 수
 	private int memberMonth;	//이달 회원수
 	
-	private int totalSales;		//이달 판매 총액
-	private int totalChat;		//이달 상담 총액
-	private int totalClass;		//이달 클래스 총액
-	private int totalIdea;		//이달 지식마켓 총액
+	private long totalSales;		//이달 판매 총액
+	private long totalChat;		//이달 상담 총액
+	private long totalClass;		//이달 클래스 총액
+	private long totalIdea;		//이달 지식마켓 총액
 }
